Error logging, ability language, vvardenfell and style motifs table
Added a error logging to database.
Abilites now also exports language it's exported in.
Added zone Vvardenfell for upcoming Morrowind.
Showing i item style now shows a table over your characters of with checks for known motifs.

+++ database/migrations/2017_04_05_210416_create_logs_table.php
@@ -15,7 +15,7 @@ class CreateLogsTable extends Migration
     {
         Schema::create('logs', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->tinyInteger('code');
+            $table->integer('code');
             $table->tinyInteger('severity')->nullable();
 
             $table->string('error', 2000);
